♻️refactor: 이미지 등록 관련 API 리팩토링
- 프로젝트에 따른 Layer 구조를 통일합니다.
- API Response 형식을 통일합니다.
