+++ backend/src/main/java/codesquard/app/api/response/ResponseMessage.java
@@ -3,5 +3,7 @@ package codesquard.app.api.response;
 public abstract class ResponseMessage {
 
 	public static final String COMMENT_SAVE_SUCCESS = "댓글 등록에 성공했습니다.";
+	public static final String IMAGE_SAVE_SUCCESS = "이미지 저장에 성공했습니다.";
+	public static final String MAXIMUM_UPLOAD_SIZE_EXCEEDED = "파일의 최대 크기를 초과했습니다.";
 
 }
