+++ backend/src/main/java/codesquard/app/image/config/S3Config.java
@@ -1,4 +1,4 @@
-package codesquard.app.image;
+package codesquard.app.image.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
